Create a recipe main page & 3 single recipe pages
Created a recipe main page to house titles of recipes and links to
pages housing those individual recipes. Each recipe page contains an
image, description, ingredient list, and directions. This is the first
step in a series to create a recipes website for the odin project's
foundations course and will be built upon going forward.

(Did the commit process in the recipes folder... committing the rest.)